// jobup-frontend/src/app/generated-sources/openapi/model/fileUploadResponseDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * File upload response DTO
 */
export interface FileUploadResponseDto { 
    /**
     * File ID
     */
    id?: string;
    /**
     * Original file name
     */
    originalFileName?: string;
    /**
     * File download URL
     */
    downloadUrl?: string;
    /**
     * File content type
     */
    contentType?: string;
    /**
     * File size in bytes
     */
    fileSize?: number;
    /**
     * File type
     */
    fileType?: FileUploadResponseDto.FileTypeEnum;
    /**
     * File category
     */
    category?: FileUploadResponseDto.CategoryEnum;
    /**
     * Upload timestamp
     */
    uploadedAt?: string;
    /**
     * updated timestamp
     */
    updatedAt?: string;
    /**
     * Image width (for images only)
     */
    width?: number;
    /**
     * Image height (for images only)
     */
    height?: number;
    /**
     * Owner ID
     */
    ownerId?: string;
}
export namespace FileUploadResponseDto {
    export type FileTypeEnum = 'IMAGE_JPEG' | 'IMAGE_PNG' | 'IMAGE_GIF' | 'DOCUMENT_PDF' | 'DOCUMENT_DOC' | 'DOCUMENT_DOCX';
    export const FileTypeEnum = {
        ImageJpeg: 'IMAGE_JPEG' as FileTypeEnum,
        ImagePng: 'IMAGE_PNG' as FileTypeEnum,
        ImageGif: 'IMAGE_GIF' as FileTypeEnum,
        DocumentPdf: 'DOCUMENT_PDF' as FileTypeEnum,
        DocumentDoc: 'DOCUMENT_DOC' as FileTypeEnum,
        DocumentDocx: 'DOCUMENT_DOCX' as FileTypeEnum
    };
    export type CategoryEnum = 'PROFILE_PICTURE' | 'JOB_ATTACHMENT' | 'WORKER_PORTFOLIO' | 'WORKER_CERTIFICATE';
    export const CategoryEnum = {
        ProfilePicture: 'PROFILE_PICTURE' as CategoryEnum,
        JobAttachment: 'JOB_ATTACHMENT' as CategoryEnum,
        WorkerPortfolio: 'WORKER_PORTFOLIO' as CategoryEnum,
        WorkerCertificate: 'WORKER_CERTIFICATE' as CategoryEnum
    };
}